fix tx ack queue contention
xmit_skb_handles must not be used anymore after clearing DataHandle.
0 means free and another thread can take it before we're ready.
Fix race condition in allocation of an xmit_skb_handles slot, when more
than 1 thread (e.g. pppd and kernel thread) try to allocate. It happened
that a confirmed Pkt Id was not in the list because it got overridden.