Fix lot of checkpatch issues
This should not happen again, please check every commit with
checkpatch --strict, this can be easy done with
git diff >somefile
/usr/src/linux/scripts/checkpatch.pl --strict somefile

Some of the echo canceler have still big issues, not only checkpatch
related, some are still architecture specific.
A other issue with the echo canceler is, that some are already mainline in
the staging area, so we need to use them and their interface, this need to
be solved before we can submit the echocanceler stuff to mainline.

Signed-off-by: Karsten Keil <keil@b1systems.de>