Add autoconf/automake support
Trying to get rid of proper standalone compiling without need
to delete files in kernel include directories.
So it should allow standalone builds even if the kernel already
has mISDN.
In the end it should allow to have only one socket source tree
(include/... drivers/...) with no differences to the in kernel
tree for Makefiles/Kconfig and so on.
The trick is, that the standalone build will use a copy of the
include/... drivers/... directories with some modifications
under standalone/.
It is working now, but still not complete.

Signed-off-by: Karsten Keil <keil@b1-systems.de>